pyproject.toml
[project]
name = "aifoundry"
version = "0.1.0"
description = "Backend de AI genérico que soporta los 4 paradigmas de AI: LLM Workflow, RAG, AI Agent y Multi-Agent"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "AIFoundry Team" }
]
keywords = ["ai", "langchain", "langgraph", "fastapi", "agents", "rag", "mcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # LangChain ecosystem
    "langchain",
    "langchain-core",
    "langgraph",
    "langchain-community",
    "langchain-openai",
    "langchain-anthropic",
    
    # FastAPI
    "fastapi",
    "uvicorn[standard]",
    
    # LLM providers
    "litellm",
    
    # MCP
    "langchain-mcp-adapters",
    "fastmcp",
    
    # Validation & Settings
    "pydantic",
    "pydantic-settings",
    
    # Scraping
    "beautifulsoup4",
    "markdownify",
    "readability-lxml",

    # Utilities
    "httpx[http2]",
    "tenacity",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "mypy",
    "pre-commit",
]

docs = [
    "mkdocs",
    "mkdocs-material",
]

[project.urls]
Homepage = "https://github.com/user/aifoundry"
Documentation = "https://github.com/user/aifoundry#readme"
Repository = "https://github.com/user/aifoundry"

[project.scripts]
aifoundry = "aifoundry.app.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["aifoundry"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["aifoundry"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["aifoundry/tests"]
filterwarnings = [
    "ignore::DeprecationWarning",
]
